Revert "fix(sns-subscriptions): SQS queue encrypted by AWS managed KMS key is allowed to be specified as subscription and dead-letter queue (#26110)"
This reverts commit 0531492451b4f99fe469380ba926f22addbfc492.
